PP-8030 refactor performance report url
- the endpoint access url and json body has changed in ldger. refactor
  to reflect those changes
