+++ app/services/create_performance_json.service.js
@@ -4,7 +4,7 @@ const moment = require('moment')
 const lodash = require('lodash')
 
 const filteredServices = require('./get_filtered_services.service')
-const performanceDataService = require('./get_performance_data.service')
+const performanceDataService = require('./get_performance_report.service')
 
 async function createPerformanceJson() {
   const fromDate = moment.utc().subtract(7, 'days').startOf('day').format()
@@ -27,8 +27,8 @@ async function createPerformanceJson() {
 
   return  {
     dateUpdated: moment().format('D MMMM YYYY'),
-    numberOfPayments: (performanceReport.payments.count / 1000000).toFixed(1) + ' million',
-    totalPaymentAmount: (performanceReport.payments.gross_amount / 1000000).toFixed(1) + ' million',
+    numberOfPayments: (performanceReport.total_volume / 1000000).toFixed(1) + ' million',
+    totalPaymentAmount: (performanceReport.total_amount / 1000000).toFixed(1) + ' million',
     numberOfServices: services.length,
     numberOfOrganisations: Object.keys(countByOrganisation).length,
     serviceCountBySector: countBySector,
