+++ app/services/get_performance_report.service.js
@@ -3,7 +3,7 @@
 const ledgerClient = require('./clients/ledger.client')
 
 async function getPerformanceReport(fromDate, toDate) {
-  return  await ledgerClient.getTransactionsSummary(fromDate, toDate)
+  return  await ledgerClient.getPerformanceReport(fromDate, toDate)
 }
 
 module.exports = {
