+++ test/integration/get_performance_report.services.it.test.js
@@ -5,7 +5,7 @@ const chai = require('chai')
 
 const { expect } = chai
 
-const performanceDataService = require('../../app/services/get_performance_data.service')
+const performanceDataService = require('../../app/services/get_performance_report.service')
 const fixtures = require('../fixtures/services.fixture')
 
 const ledgerMock = nock(process.env.LEDGER_URL)
@@ -19,15 +19,15 @@ describe('get performance data', () => {
   it('should return performance data from ledger', async function () {
     const fromDate =  '2021-04-27T00:00:00Z'
     const toDate = '2021-05-04T00:00:00Z'
-    const url = '/v1/report/transactions-summary?override_account_id_restriction=true&from_date=' + fromDate + '&to_date=' + toDate
+    const url = '/v1/report/performance-report?from_date=' + fromDate + '&to_date=' + toDate
 
     ledgerMock.get(url)
       .reply(200, fixtures.statisticsData())
 
     const data = await performanceDataService.getPerformanceReport(fromDate, toDate)
 
-    expect(data.payments.count).to.equal(4949300)
-    expect(data.payments.gross_amount).to.equal(2623170000)
-    expect(data.net_income).to.equal(26021000000)
+    expect(data.total_volume).to.equal(4949300)
+    expect(data.total_amount).to.equal(2623170000)
+    expect(data.average_amount).to.equal(5100)
   })
 })
